Mpesa Integration
Integrated the Lipa na Mpesa API however there is an issue yet to be resolved. See issues for reference

+++ app/src/main/java/com/example/shopit/data/model/LipaNaMpesaRequest.kt
@@ -9,7 +9,7 @@ data class LipaNaMpesaRequest(
     @Json(name = "AccountReference")
     var accountReference: String,
     @Json(name = "Amount")
-    var amount: Int,
+    var amount: Double,
     @Json(name = "BusinessShortCode")
     var businessShortCode: Int,
     @Json(name = "CallBackURL")
